WIP openshift appliance deployment model
Using openshift appliance, we can embed grout and frr container images.  This
raw file can be either installed in factory or streamed with the BMC interface.
This can be very slow on a 1Gb BMC interface (4h+).

I have to test with a 10Gb bmc interface, but this method applies to any agent
based installer, including SNO.